+++ daft/daft.pyi
@@ -1047,8 +1047,6 @@ class PyExpr:
     def mean(self) -> PyExpr: ...
     def min(self) -> PyExpr: ...
     def max(self) -> PyExpr: ...
-    def shift_left(self, bits: PyExpr) -> PyExpr: ...
-    def shift_right(self, bits: PyExpr) -> PyExpr: ...
     def hash(self, seed: Any | None = None) -> PyExpr: ...
     def minhash(
         self,
@@ -1213,8 +1211,6 @@ class PySeries:
     def mean(self) -> PySeries: ...
     def min(self) -> PySeries: ...
     def max(self) -> PySeries: ...
-    def shift_left(self, bits: PySeries) -> PySeries: ...
-    def shift_right(self, bits: PySeries) -> PySeries: ...
     def agg_list(self) -> PySeries: ...
     def cast(self, dtype: PyDataType) -> PySeries: ...
     def ceil(self) -> PySeries: ...
