+++ src/daft-core/src/datatypes/dtype.rs
@@ -287,14 +287,6 @@ impl DataType {
     }
 
     #[inline]
-    pub fn is_unsigned(&self) -> bool {
-        matches!(
-            self,
-            DataType::UInt8 | DataType::UInt16 | DataType::UInt32 | DataType::UInt64
-        )
-    }
-
-    #[inline]
     pub fn is_floating(&self) -> bool {
         matches!(
             self,
