+++ src/daft-dsl/src/functions/numeric/mod.rs
@@ -4,7 +4,6 @@ mod exp;
 mod floor;
 mod log;
 mod round;
-mod shift;
 mod sign;
 mod sqrt;
 mod trigonometry;
@@ -50,8 +49,6 @@ pub enum NumericExpr {
     Log(FloatWrapper<f64>),
     Ln,
     Exp,
-    ShiftRight,
-    ShiftLeft,
 }
 
 impl NumericExpr {
@@ -79,8 +76,6 @@ impl NumericExpr {
             Log10 => &LogEvaluator(log::LogFunction::Log10),
             Log(_) => &LogEvaluator(log::LogFunction::Log),
             Ln => &LogEvaluator(log::LogFunction::Ln),
-            ShiftLeft => &shift::ShiftEvaluator(shift::ShiftFunction::Left),
-            ShiftRight => &shift::ShiftEvaluator(shift::ShiftFunction::Right),
             Exp => &ExpEvaluator {},
         }
     }
@@ -242,22 +237,6 @@ pub fn ln(input: ExprRef) -> ExprRef {
     Expr::Function {
         func: super::FunctionExpr::Numeric(NumericExpr::Ln),
         inputs: vec![input],
-    }
-    .into()
-}
-
-pub fn shift_left(input: ExprRef, shift: ExprRef) -> ExprRef {
-    Expr::Function {
-        func: super::FunctionExpr::Numeric(NumericExpr::ShiftLeft),
-        inputs: vec![input, shift],
-    }
-    .into()
-}
-
-pub fn shift_right(input: ExprRef, shift: ExprRef) -> ExprRef {
-    Expr::Function {
-        func: super::FunctionExpr::Numeric(NumericExpr::ShiftRight),
-        inputs: vec![input, shift],
     }
     .into()
 }
